🚚 move: Footer device 폴더 분리
