+++ app/src/components/elements/Footer/Mobile/index.tsx
