+++ app/src/components/elements/Footer/index.tsx
@@ -1,6 +1,6 @@
 import { TabletAndAbove, Mobile } from '@utils/responsive/Device';
-import { TabletAndAboveFooter } from './TabletAndAboveFooter';
-import { MobileFooter } from './MobileFooter';
+import { TabletAndAboveFooter } from './TabletAndAbove';
+import { MobileFooter } from './Mobile';
 
 export const Footer = () => {
   return (
